requirements.txt
wordcloud>=1.9.0
matplotlib>=3.5.0
numpy>=1.21.0
Pillow>=8.0.0 